+++ src/main/java/net/geforcemods/securitycraft/blockentities/BlockChangeDetectorBlockEntity.java
@@ -33,7 +33,7 @@ import net.minecraft.util.text.ITextComponent;
 import net.minecraftforge.common.util.Constants;
 
 public class BlockChangeDetectorBlockEntity extends DisguisableBlockEntity implements IInventory, INamedContainerProvider, ILockable, ITickableTileEntity {
-	private IntOption signalLength = new IntOption(this::getBlockPos, "signalLength", 60, 5, 400, 5, true); //20 seconds max
+	private IntOption signalLength = new IntOption(this::getBlockPos, "signalLength", 60, 0, 400, 5, true); //20 seconds max
 	private IntOption range = new IntOption(this::getBlockPos, "range", 5, 1, 15, 1, true);
 	private DisabledOption disabled = new DisabledOption(false);
 	private IgnoreOwnerOption ignoreOwner = new IgnoreOwnerOption(true);
@@ -67,9 +67,13 @@ public class BlockChangeDetectorBlockEntity extends DisguisableBlockEntity imple
 			return;
 
 		if (isModuleEnabled(ModuleType.REDSTONE)) {
-			level.setBlockAndUpdate(worldPosition, getBlockState().setValue(BlockChangeDetectorBlock.POWERED, true));
+			int signalLength = getSignalLength();
+
+			level.setBlockAndUpdate(worldPosition, getBlockState().cycle(BlockChangeDetectorBlock.POWERED));
 			BlockUtils.updateIndirectNeighbors(level, worldPosition, SCContent.BLOCK_CHANGE_DETECTOR.get());
-			level.getBlockTicks().scheduleTick(worldPosition, SCContent.BLOCK_CHANGE_DETECTOR.get(), signalLength.get());
+
+			if (signalLength > 0)
+				level.getBlockTicks().scheduleTick(worldPosition, SCContent.BLOCK_CHANGE_DETECTOR.get(), signalLength);
 		}
 
 		entries.add(new ChangeEntry(player.getDisplayName().getString(), player.getUUID(), System.currentTimeMillis(), action, pos, state));
@@ -143,6 +147,10 @@ public class BlockChangeDetectorBlockEntity extends DisguisableBlockEntity imple
 
 	public DetectionMode getMode() {
 		return mode;
+	}
+
+	public int getSignalLength() {
+		return signalLength.get();
 	}
 
 	public int getRange() {