+++ src/main/java/net/geforcemods/securitycraft/blockentities/LaserBlockBlockEntity.java
@@ -72,6 +72,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 	});
 	private LazyOptional<IItemHandler> insertOnlyHandler, lensHandler;
 	private LensContainer lenses = new LensContainer(6);
+	private long lastToggleTime;
 
 	public LaserBlockBlockEntity() {
 		super(SCContent.LASER_BLOCK_BLOCK_ENTITY.get());
@@ -133,8 +134,10 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 			}
 			else if (option.getName().equals("ignoreOwner"))
 				ignoreOwner.copy(option);
-			else if (option.getName().equals("signalLength"))
+			else if (option.getName().equals("signalLength")) {
 				signalLength.copy(option);
+				turnOffRedstoneOutput();
+			}
 		}
 		else if (action instanceof ILinkedAction.ModuleInserted) {
 			ILinkedAction.ModuleInserted moduleInserted = (ILinkedAction.ModuleInserted) action;
@@ -154,15 +157,32 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 		else if (action instanceof ILinkedAction.StateChanged<?>) {
 			BlockState state = getBlockState();
 
-			if (((ILinkedAction.StateChanged<?>) action).property == LaserBlock.POWERED && !state.getValue(LaserBlock.POWERED)) {
-				level.setBlockAndUpdate(worldPosition, state.setValue(LaserBlock.POWERED, true));
+			if (((ILinkedAction.StateChanged<?>) action).property == LaserBlock.POWERED) {
+				int signalLength = getSignalLength();
+
+				level.setBlockAndUpdate(worldPosition, state.cycle(LaserBlock.POWERED));
 				BlockUtils.updateIndirectNeighbors(level, worldPosition, SCContent.LASER_BLOCK.get());
-				level.getBlockTicks().scheduleTick(worldPosition, SCContent.LASER_BLOCK.get(), signalLength.get());
+
+				if (signalLength > 0)
+					level.getBlockTicks().scheduleTick(worldPosition, SCContent.LASER_BLOCK.get(), signalLength);
 			}
 		}
 
 		excludedBEs.add(this);
 		createLinkedBlockAction(action, excludedBEs);
+	}
+
+	@Override
+	public void onOptionChanged(Option<?> option) {
+		if (option.getName().equals(signalLength.getName()))
+			turnOffRedstoneOutput();
+
+		super.onOptionChanged(option);
+	}
+
+	private void turnOffRedstoneOutput() {
+		level.setBlockAndUpdate(worldPosition, getBlockState().setValue(LaserBlock.POWERED, false));
+		BlockUtils.updateIndirectNeighbors(level, worldPosition, getBlockState().getBlock());
 	}
 
 	@Override
@@ -182,10 +202,8 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 		if (module == ModuleType.DISGUISE)
 			DisguisableBlockEntity.onDisguiseModuleRemoved(this, stack, toggled);
 		else if (module == ModuleType.REDSTONE) {
-			if (getBlockState().getValue(LaserBlock.POWERED)) {
-				level.setBlockAndUpdate(worldPosition, getBlockState().setValue(LaserBlock.POWERED, false));
-				BlockUtils.updateIndirectNeighbors(level, worldPosition, SCContent.LASER_BLOCK.get());
-			}
+			if (getBlockState().getValue(LaserBlock.POWERED))
+				turnOffRedstoneOutput();
 		}
 		else if (module == ModuleType.SMART)
 			applyExistingSideConfig();
@@ -331,6 +349,18 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 
 	public int getSignalLength() {
 		return signalLength.get();
+	}
+
+	public void setLastToggleTime(long lastToggleTime) {
+		this.lastToggleTime = lastToggleTime;
+	}
+
+	public long getLastToggleTime() {
+		return lastToggleTime;
+	}
+
+	public long timeSinceLastToggle() {
+		return System.currentTimeMillis() - getLastToggleTime();
 	}
 
 	public void applyNewSideConfig(Map<Direction, Boolean> sideConfig, PlayerEntity player) {