+++ src/main/java/net/geforcemods/securitycraft/blockentities/RiftStabilizerBlockEntity.java
@@ -44,7 +44,7 @@ import net.minecraftforge.event.entity.living.EntityTeleportEvent.SpreadPlayersC
 import net.minecraftforge.event.entity.living.EntityTeleportEvent.TeleportCommand;
 
 public class RiftStabilizerBlockEntity extends DisguisableBlockEntity implements ITickableTileEntity, ILockable, IToggleableEntries<TeleportationType> {
-	private final IntOption signalLength = new IntOption(this::getBlockPos, "signalLength", 60, 5, 400, 5, true); //20 seconds max
+	private final IntOption signalLength = new IntOption(this::getBlockPos, "signalLength", 60, 0, 400, 5, true); //20 seconds max
 	private final IntOption range = new IntOption(this::getBlockPos, "range", 5, 1, 15, 1, true);
 	private final DisabledOption disabled = new DisabledOption(false);
 	private IgnoreOwnerOption ignoreOwner = new IgnoreOwnerOption(true);