+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockChangeDetectorBlock.java
@@ -86,8 +86,10 @@ public class BlockChangeDetectorBlock extends DisguisableBlock {
 
 	@Override
 	public void tick(BlockState state, ServerWorld level, BlockPos pos, Random random) {
-		level.setBlockAndUpdate(pos, state.setValue(POWERED, false));
-		BlockUtils.updateIndirectNeighbors(level, pos, this);
+		if (state.getValue(POWERED)) {
+			level.setBlockAndUpdate(pos, state.setValue(POWERED, false));
+			BlockUtils.updateIndirectNeighbors(level, pos, this);
+		}
 	}
 
 	@Override