+++ Changelog.md
@@ -12,6 +12,7 @@
 - New: SecurityCraft's /sc command can now be accessed via /securitycraft as well
 - New: "/sc owner" command to set/fill the owner of blocks
 - New: The Bouncing Betty, Claymore, Frame, Mine, Panic Button, and Security Camera are now waterloggable
+- New: Several blocks with a signal length option for redstone signal length can now be set to toggle the redstone signal instead by setting the option to 0
 - Change: Sentries now shoot half as frequently to match the shooting rate in newer Minecraft versions
 - Change: The Reinforced Hopper screen now shows "Reinforced Hopper" instead of "Item Hopper" as its default inventory title
 - Change: The hitbox of the Sentry has been changed, which means that only the base of Sentries can be interacted with now
@@ -46,6 +47,8 @@
 - Fix: Mobs can be converted multiple times by an Electrified Iron Fence (Gate)
 - Fix: Incorrect button is being hidden while navigating in the SecurityCraft Manual
 - Fix: Anyone can convert a Frame/Reinforced Iron Trapdoor to a Keypad/Passcode-protected Trapdoor
+- Fix: The Keypad Trapdoor's signal length option cannot be set to 0 despite the description saying it can
+- Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
 
 --------------------------Changelog for v1.9.8 of SecurityCraft--------------------------